refactor: move filter item to separate component
